debugging part 34

+++ backend/db/models/spot.js
@@ -64,7 +64,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     lat: {
-      type: DataTypes.DECIMAL(10,7),
+      type: DataTypes.FLOAT,
       allowNull: false,
       validate: {
         min: -90,
@@ -72,7 +72,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     lng: {
-      type: DataTypes.DECIMAL(10,7),
+      type: DataTypes.FLOAT,
       allowNull: false,
       validate: {
         min: -180,
@@ -94,7 +94,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     price: {
-      type: DataTypes.DECIMAL(7,2),
+      type: DataTypes.FLOAT,
       allowNull: false,
       validate: {
         notEmpty: true,
